+++ NEWS.md
@@ -1,6 +1,6 @@
 # eeguana 0.1.3.9000
   - Changes:
-    - `read_` functions get a `samples` arguments to read a segment of samples from the file. 
+    - 
   - Bugs:
     - `read_edf()` wasn't reading events from the status channel
     - other minor bugs